+++ NEWS.md
@@ -1,17 +1,17 @@
 # fpp3 (development version)
 
-  * Added bank_calls, prices and souvenirs data sets
+  * Added bank_calls, prices, souvenirs and insurance data sets
 
 # fpp3 (0.3)
  
   * Updated all objects to work with tsibble v0.9.0
   * Added series titles to us_employment
 
-# fpp3 0.2
+# fpp3 (0.2)
 
  * Added aus_arrivals data set.
 
-# fpp3 0.1
+# fpp3 (0.1)
 
  * Initial version based on the tidyverse package by Hadley Wickham.
  * Some data ported from fpp2. New data sets include us_employment.